# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Uint8Array": {
				"description": "A typed array of 8-bit unsigned integer values. The contents are initialized to 0. If the\r\nrequested number of bytes could not be allocated an exception is raised.",
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"CommitmentScriptDataType": {
				"properties": {
					"tweakedPubkey": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"tapMerkleRoot": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"tapLeafScript": {},
					"tapInternalKey": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"leaves": {},
					"wsh": {
						"type": "string"
					},
					"witnessScript": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"redeemScript": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"paymentType": {
						"type": "string"
					},
					"script": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"address": {
						"type": "string"
					}
				},
				"required": [
					"paymentType"
				],
				"type": "object"
			},
			"VoutI": {
				"properties": {
					"value": {
						"type": "number",
						"format": "double"
					},
					"scriptpubkey_address": {
						"type": "string"
					},
					"scriptpubkey_type": {
						"type": "string"
					},
					"scriptpubkey_asm": {
						"type": "string"
					},
					"scriptpubkey": {
						"type": "string"
					}
				},
				"required": [
					"value",
					"scriptpubkey_address",
					"scriptpubkey_type",
					"scriptpubkey_asm",
					"scriptpubkey"
				],
				"type": "object"
			},
			"CommitmentStatus": {
				"enum": [
					0,
					1,
					2,
					3,
					4
				],
				"type": "number"
			},
			"RevealerTxModes": {
				"enum": [
					"OP_RETURN",
					"OP_DROP"
				],
				"type": "string"
			},
			"RevealerTxTypes": {
				"enum": [
					"SBTC_DEPOSIT",
					"SBTC_WITHDRAWAL"
				],
				"type": "string"
			},
			"RevealerTransaction": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"txId": {
						"type": "string"
					},
					"psbt": {
						"type": "string"
					},
					"originator": {
						"type": "string"
					},
					"commitment": {
						"$ref": "#/components/schemas/CommitmentScriptDataType"
					},
					"vout": {
						"$ref": "#/components/schemas/VoutI"
					},
					"signed": {
						"type": "boolean"
					},
					"recipient": {
						"type": "string"
					},
					"amountSats": {
						"type": "number",
						"format": "double"
					},
					"confirmations": {
						"type": "number",
						"format": "double"
					},
					"created": {
						"type": "number",
						"format": "double"
					},
					"updated": {
						"type": "number",
						"format": "double"
					},
					"signature": {
						"type": "string"
					},
					"paymentPublicKey": {
						"type": "string"
					},
					"paymentAddress": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/CommitmentStatus"
					},
					"mode": {
						"$ref": "#/components/schemas/RevealerTxModes"
					},
					"type": {
						"$ref": "#/components/schemas/RevealerTxTypes"
					},
					"blockHeight": {
						"type": "number",
						"format": "double"
					},
					"sbtcPublicKey": {
						"type": "string"
					}
				},
				"required": [
					"txId",
					"originator",
					"signed",
					"recipient",
					"amountSats",
					"confirmations",
					"created",
					"updated",
					"paymentPublicKey",
					"paymentAddress",
					"status",
					"mode",
					"type",
					"blockHeight",
					"sbtcPublicKey"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "revealer-api",
		"version": "1.0.0",
		"description": "Proxy server for ordicall web application to access ordinals and bitcoin servers",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/revealer-api/v1/transactions/scan-sbtc-wallet-transaction/{txid}": {
			"get": {
				"operationId": "ScanSbtcWalletTransaction",
				"responses": {
					"200": {
						"description": "array of transactions",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RevealerTransaction"
								}
							}
						}
					}
				},
				"description": "Fetch historic revealer transactions from an sbtc wallet - does not overwrite existing txs.",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "txid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/revealer-api/v1/transactions/scan-sbtc-wallet-transactions/{address}": {
			"get": {
				"operationId": "ScanSbtcWalletTransactions",
				"responses": {
					"200": {
						"description": "array of transactions",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Fetch historic revealer transactions from an sbtc wallet - does not overwrite existing txs.",
				"security": [],
				"parameters": [
					{
						"description": "address of the sbtc wallet to scan",
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/revealer-api/v1/transactions/scan-commitment-transactions/{address}": {
			"post": {
				"operationId": "ScanRevealerTransactionsByCommitAddress",
				"responses": {
					"200": {
						"description": "array of transactions",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Fetch transactions at the given address - we only ever expect one - the users deposit.",
				"security": [],
				"parameters": [
					{
						"description": "address of the commitment",
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/revealer-api/v1/transactions/check-revealer-transaction/{address}": {
			"get": {
				"operationId": "CheckRevealerTransactionByCommitAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/revealer-api/v1/transactions/connect-revealer-transaction/{address}/{txId}": {
			"get": {
				"operationId": "ConnectRevealerTransactionByCommitAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "txId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/revealer-api/v1/transactions/get-revealer-transactions/{page}/{limit}": {
			"get": {
				"operationId": "GetRevealerTransactions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/revealer-api/v1/transactions/get-revealer-transactions-by-commit-address/{address}": {
			"get": {
				"operationId": "GetRevealerTransactionByCommitAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/revealer-api/v1/transactions/get-revealer-transactions-by-txid/{txId}": {
			"get": {
				"operationId": "GetRevealerTransactionByTxId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "txId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/revealer-api/v1/transactions/get-revealer-transactions-by-originator/{address}/{page}/{limit}": {
			"get": {
				"operationId": "GetRevealerTransactionByOriginator",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "page",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "limit",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/revealer-api/v1/transactions/get-revealer-transactions-pending-by-originator/{address}": {
			"get": {
				"operationId": "GetRevealerTransactionPendingByOriginator",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}